Write a new ipfs/client.go that mimis the functionality of the old http client.
